Creation of a WeatherUtils object
